Refactor task-related functions out of main database function
